+++ ananse/commands/__init__.py
@@ -1,5 +1,5 @@
 from ananse.commands.binding import binding  # noqa
 from ananse.commands.influence import influence  # noqa
 from ananse.commands.network import network  # noqa
-from ananse.commands.view import view  # noqa
 from ananse.commands.plot import plot  # noqa
+from ananse.commands.view import view  # noqa