+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.5.0] - 2023-10-15
+
+### Added
+- `ananse influence` now accepts a whitelist of genes and/or interactions from the network
+  - these will be added back into the influence network after selecting top edges.
+
+### Fixed
+- No region overlap (pandas str.replace() no longer uses regex by default)
+- Error while merging duplicate genes in ANANSE binding
+
 ## [0.4.1] - 2023-01-12
 
 ### Added